refactor: 체크 박스 컴포넌트 패키지 및 이름 수정
